OpenVMS: c wrapper building and test programs minor fixes
git-svn-id: http://svn.fastmq.org/svn/zeromq/trunk@1225 7451a71c-b6c8-49cc-8254-141fc5361860

+++ perf/tests/zmq/c_local_lat.c
@@ -83,7 +83,7 @@ int main (int argc, char *argv [])
     int counter;
     void *out_buf;
     void *in_buf;
-    size_t in_size;
+    uint64_t in_size;
     uint64_t start;
     uint64_t end; 
     double latency;
